+++ src/main/java/de/maxhenkel/audioplayer/api/AudioPlayerApi.java
@@ -79,4 +79,6 @@ public interface AudioPlayerApi {
 
     Collection<AudioFileMetadata> getAllMetadata();
 
+    long getMaxUploadSize();
+
 }