+++ changelog.md
@@ -1,3 +1,4 @@
 - Made import API return a future
 - Added audio file metadata access to the API
-- Added a way to list all audio file metadata
+- Added a way to list all audio file metadata with the API
+- Added a way to get the maximum upload size with the API