+++ changelog.md
@@ -2,3 +2,4 @@
 - Added audio file metadata access to the API
 - Added a way to list all audio file metadata with the API
 - Added a way to get the maximum upload size with the API
+- Added a way to get upload folder with the API