+++ changelog.md
@@ -3,3 +3,4 @@
 - Added a way to list all audio file metadata with the API
 - Added a way to get the maximum upload size with the API
 - Added a way to get upload folder with the API
+- Allow silently importing audio as a player using the API