+++ changelog.md
@@ -8,3 +8,4 @@
 - Allow applying new audio data to an item
 - Fixed serverfile suggestions suggesting folders
 - Added rename command
+- Added rename function to the API