+++ changelog.md
@@ -5,3 +5,4 @@
 - Removed file extensions from audio file names
 - Made audio file names unique
 - Create backup of metadata before upgrade
+- Allow applying new audio data to an item