+++ changelog.md
@@ -7,3 +7,4 @@
 - Create backup of metadata before upgrade
 - Allow applying new audio data to an item
 - Fixed serverfile suggestions suggesting folders
+- Added rename command