+++ changelog.md
@@ -1,2 +1,3 @@
 - Switched to fabric-permission-api-v1
 - Removed Lucko's fabric-permissions-api
+- Updated dependencies of the web upload frontend