+++ changelog.md
@@ -1,3 +1,4 @@
 - Switched to fabric-permission-api-v1
 - Removed Lucko's fabric-permissions-api
 - Updated dependencies of the web upload frontend
+- Updated to 26.3